add status text line
This makes it possible to add a "Pumpkin Power" text label above
the catapult charge-up indicator bar. Hopefully this will make it
obvious in video or screenshots what's going on with the pumpkin
charge-up mechanic.

It's easy to understand when and why the charge-up bar appears if
you're the one holding the gamepad and shooting pumpkins. But, from
just watching a video of gameplay, there's a lot less context about
how the charge-up bar relates to gamepad input. Adding the text
label should make it less confusing to follow along on video.