lots of comments for code.py::main()
Trying to make it so reading through the comments in main() will
give people a chance to understand more about my design decisions.
I've tried a lot of stuff along the way to my current approach, and
much of that is non-obvious. So, I'm trying to write down more
context and rationale.