show USB status on display's status line
This makes it easier to follow what's going on with the USB gamepad
connection status by looking only at the TFT display without the
need to check the serial console.

This will make it more obvious in videos that a USB gamepad is
being used. It should also help make it less confusing IRL when
you try to press START before the gamepad is ready after a reset
(I tend to do that).